// app/src/types.ts
// Type definition for preference keys
export type PreferenceKeys =
  | 'weather_code'
  | 'temperature_2m_max'
  | 'temperature_2m_min'
  | 'apparent_temperature_max'
  | 'apparent_temperature_min'
  | 'sunrise'
  | 'sunset'
  | 'daylight_duration'
  | 'sunshine_duration'
  | 'uv_index_max'
  | 'precipitation_sum'
  | 'rain_sum'
  | 'showers_sum'
  | 'snowfall_sum'
  | 'precipitation_hours'
  | 'precipitation_probability_max'
  | 'wind_speed_10m_max'
  | 'wind_gusts_10m_max'
  | 'wind_direction_10m_dominant';